split binary format test into individual files for each structure
