+++ __tests__/binary/constants/externalAttrs.ts
@@ -51,7 +51,7 @@ export const UNIX_TYPES = {
  */
 // prettier-ignore
 export const UNIX_PERMISSIONS = {
-  USER_READ: 0o400, USER_WRITE: 0o200, USER_EXEC: 0o100,
+  USER_READ:  0o400, USER_WRITE:  0o200, USER_EXEC:  0o100,
   GROUP_READ: 0o040, GROUP_WRITE: 0o020, GROUP_EXEC: 0o010,
   OTHER_READ: 0o004, OTHER_WRITE: 0o002, OTHER_EXEC: 0o001,
 } as const;
